moves around some code
Removes configuration into it's own module, also moves request related
functions into it's own `Contentful.Request` module.
