+++ test/contentful/misc_test.exs
@@ -1,0 +1,5 @@
+defmodule Contentful.MiscTest do
+  use ExUnit.Case
+
+  doctest Contentful.Misc
+end
